requirements.txt
Django==2.1.2
djangorestframework==3.8.2
mpu==0.15.0
pkg-resources==0.0.0
pytz==2018.5
